01-applications/fss-dataxflow-backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<0.35
pydantic-settings>=2.7,<3.0
PyJWT>=2.10,<3.0
kubernetes>=31.0,<33.0
pymongo>=4.10,<5.0
redis>=5.2,<6.0
teradatasql>=20.0
psycopg[binary]>=3.2,<4.0
SQLAlchemy>=2.0,<3.0
